configure.ac: remove `-flat_namespace` usage
Apple discourages the use of the `-flat_namespace` flag. It exists only
for compatibility with very old OS X versions and is considered
effectively deprecated according to
https://developer.apple.com/forums/thread/689991. Using the flag can
cause linker errors (e.g. due to name collisions), so it is best to
avoid usage of the flag unless required.

This commit also changes `-undefined suppress` to `-undefined dynamic_lookup`,
so that undefined symbols will be looked up at runtime.