Alt + Tab has never worked to set previous palette
Alt + Tab was never the key combination to set previous palette in the
graphic spectrum analyser, even if the help says so since
Cubic Player v1.0 (1995-05-29).

As a side note Shift + Tab had that function though since
Cubic Player v0.95 (1995-02-02) and was correctly documented there,
but was later replaced in Cubic Player v1.0 to set next palette for
the normal spectrum analysers.